Handling of incomplete read() operation.
As described by man:

read() attempts to read up to count bytes from file descriptor fd
       into the buffer starting at buf.

So it is possible that you will receive, in a reading operation,
less bytes that the ones you are requested. In order to complete
the read() operation you have to perform a cycle to get all the
requested bytes.

Signed-off-by: Alessandro Biasci <alessandro.biasci@huawei.com>